API unit tests: new tests

+++ test/index.mjs
@@ -16,7 +16,12 @@ import testUtils from './testUtils'
 import console from './consoleColor'
 
 const config = {
-    RESTORE_DB: 0,
+    RUN: {
+        RESTORE_DB: 1,
+        LAUNCH_API: 0
+    },
+    
+    // settings
     DB_NAME: 'graaly-test',
     //SERVER_URL: 'https://localhost:3000',
     SERVER_HOST: 'localhost',
@@ -31,31 +36,33 @@ main()
     .catch((err) => { console.error('Process stopped with error', err.stack) })
 
 async function main () {
-    if (config.RESTORE_DB) {
+    let serverProcess
+    
+    if (config.RUN.RESTORE_DB) {
         console.log('Restoring clean test DB...')
         execSync('mongorestore --drop --gzip --archive=' + config.DUMP_FILE)
         console.success('DB Restored')
-    } else {
-        console.log('Skipping DB restore (config.RESTORE_DB is falsy)')
     }
     
-    console.log('Starting API server...')
+    if (config.RUN.LAUNCH_API) {
+        console.log('Starting API server...')
     
-    let serverProcess = spawn('node', ['server.js', '--env=test'], { cwd: config.SERVER_PATH, stdio: 'inherit' })
-    
-    serverProcess = testUtils.addLoggingToProcess(serverProcess, 'API server') // default logging from spawned processes is very low
-    
-    if (!serverProcess) {
-        throw new Error('Could not start API server.')
+        serverProcess = spawn('node', ['server.js', '--env=test'], { cwd: config.SERVER_PATH, stdio: 'inherit' })
+        
+        serverProcess = testUtils.addLoggingToProcess(serverProcess, 'API server') // default logging from spawned processes is very low
+        
+        if (!serverProcess) {
+            throw new Error('Could not start API server.')
+        }
+        
+        await testUtils.sleep(2000) // let server completely startup
+        console.success('API server is running')
     }
-    
-    await testUtils.sleep(2000) // let server completely startup
-    console.success('API server is running')
     
     console.log('Run Jest test suite')
     execSync('npm run test:unit', { stdio: 'inherit' })
     
-    if (serverProcess) {
+    if (config.RUN.LAUNCH_API && serverProcess) {
         console.log('Stop API server')
         serverProcess.kill('SIGTERM')
     }
